fix: address PR review - use cw-orch root() and extract FuzzPool to pool.rs
- Use cw-orch's built-in .root() for AnyError variants instead of
  Debug-string parsing, with source() chain fallback for non-AnyError
  variants where .root() would panic
- Move FuzzPool trait to its own pool.rs file for better organization
